+++ src/client/src/commons/form/borehole/boreholeForm.js
@@ -26,7 +26,8 @@ import GroundwaterLevelMeasurement from "./hydrogeology/groundwaterLevelMeasurem
 import Hydrotest from "./hydrogeology/hydrotest";
 import { AlertContext } from "../../alert/alertContext";
 import FieldMeasurement from "./hydrogeology/fieldMeasurement";
-import ChronostratigraphyPanel from "./chronostratigraphy/chronostratigraphyPanel";
+import ChronostratigraphyPanel from "./stratigraphy/chronostratigraphyPanel";
+import LithostratigraphyPanel from "./stratigraphy/lithostratigraphyPanel";
 
 class BoreholeForm extends React.Component {
   static contextType = AlertContext;
@@ -369,7 +370,20 @@ class BoreholeForm extends React.Component {
             render={() => (
               <ChronostratigraphyPanel
                 id={parseInt(this.props.match.params.id, 10)}
-                unlocked={isEditable}
+                isEditable={isEditable}
+              />
+            )}
+          />
+          <Route
+            exact
+            path={
+              process.env.PUBLIC_URL +
+              "/editor/:id/stratigraphy/lithostratigraphy"
+            }
+            render={() => (
+              <LithostratigraphyPanel
+                id={parseInt(this.props.match.params.id, 10)}
+                isEditable={isEditable}
               />
             )}
           />